secure-flask-auth0-azure/requirements.txt
flask
python-dotenv
authlib
